Revert "Roll up master to latest tensorflow"

+++ tensorflow/core/common_runtime/bfc_allocator.h
@@ -78,7 +78,7 @@ class BFCAllocator : public VisitableAllocator {
 
   // A ChunkHandle is an index into the chunks_ vector in BFCAllocator
   // kInvalidChunkHandle means an invalid chunk
-  typedef size_t ChunkHandle;
+  typedef int ChunkHandle;
   static const int kInvalidChunkHandle = -1;
 
   typedef int BinNum;
